Reorganize the folder
